requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
idna==3.4
python-sonarqube-api==1.3.6
requests==2.28.2
tqdm==4.65.0
urllib3==1.26.15